[add/#58] PurchaseContract 추가
